add subscribe and publish on mqtt class

+++ temperature-monitoring-subsystem(esp)/src/components/api/TemperatureSensor.h
@@ -6,7 +6,7 @@
 class TemperatureSensor {
 public:
     TemperatureSensor(int pin);
-    float readTemperature();
+    int readTemperature();
 
 private:
     int sensorPin;
