+++ temperature-monitoring-subsystem(esp)/src/components/impl/TemperatureSensor.cpp
@@ -4,9 +4,9 @@ TemperatureSensor::TemperatureSensor(int pin) : sensorPin(pin) {
     pinMode(sensorPin, INPUT);
 }
 
-float TemperatureSensor::readTemperature() {
+int TemperatureSensor::readTemperature() {
     int rawValue = analogRead(sensorPin); // Lettura grezza dal sensore
-    float voltage = rawValue * (3.3 / 4095.0); // Conversione a tensione (esempio per ESP32)
-    float temperature = (voltage - 0.5) * 100; // Conversione in gradi Celsius
+    int voltage = rawValue * (3.3 / 4095.0); // Conversione a tensione (esempio per ESP32)
+    int temperature = (voltage - 0.5) * 100; // Conversione in gradi Celsius
     return temperature;
 }
